Move cache implementation to its own package
This moves the "cache" (looking less like a cache every day) into its
own package, with the memcached implementation a subpackage therein.
