+++ registry/cache/memcached/memcached.go
@@ -1,11 +1,10 @@
-package cache
+package memcached
 
 import (
 	"encoding/binary"
 	"fmt"
 	"net"
 	"sort"
-	"strings"
 	"sync"
 	"time"
 
@@ -13,42 +12,12 @@ import (
 	"github.com/go-kit/kit/log"
 	"github.com/pkg/errors"
 
-	fluxerr "github.com/weaveworks/flux/errors"
-	"github.com/weaveworks/flux/image"
+	"github.com/weaveworks/flux/registry/cache"
 )
 
 const (
 	expiry = time.Hour
 )
-
-var (
-	ErrNotCached = &fluxerr.Error{
-		Type: fluxerr.Missing,
-		Err:  memcache.ErrCacheMiss,
-		Help: `Image not yet cached
-
-It takes time to initially cache all the images. Please wait.
-
-If you have waited for a long time, check the flux logs. Potential
-reasons for the error are: no internet, no cache, error with the remote
-repository.
-`,
-	}
-)
-
-type Reader interface {
-	GetKey(k Keyer) ([]byte, time.Time, error)
-}
-
-type Writer interface {
-	SetKey(k Keyer, v []byte) error
-}
-
-type Client interface {
-	Reader
-	Writer
-	Stop()
-}
 
 // MemcacheClient is a memcache client that gets its server list from SRV
 // records, and periodically updates that ServerList.
@@ -73,7 +42,7 @@ type MemcacheConfig struct {
 	MaxIdleConns   int
 }
 
-func NewMemcacheClient(config MemcacheConfig) Client {
+func NewMemcacheClient(config MemcacheConfig) cache.Client {
 	var servers memcache.ServerList
 	client := memcache.NewFromSelector(&servers)
 	client.Timeout = config.Timeout
@@ -99,7 +68,7 @@ func NewMemcacheClient(config MemcacheConfig) Client {
 }
 
 // Does not use DNS, accepts static list of servers.
-func NewFixedServerMemcacheClient(config MemcacheConfig, addresses ...string) Client {
+func NewFixedServerMemcacheClient(config MemcacheConfig, addresses ...string) cache.Client {
 	var servers memcache.ServerList
 	servers.SetServers(addresses...)
 	client := memcache.NewFromSelector(&servers)
@@ -114,16 +83,15 @@ func NewFixedServerMemcacheClient(config MemcacheConfig, addresses ...string) Cl
 		quit:       make(chan struct{}),
 	}
 
-	client.FlushAll()
 	return newClient
 }
 
-func (c *memcacheClient) GetKey(k Keyer) ([]byte, time.Time, error) {
+func (c *memcacheClient) GetKey(k cache.Keyer) ([]byte, time.Time, error) {
 	cacheItem, err := c.Get(k.Key())
 	if err != nil {
 		if err == memcache.ErrCacheMiss {
 			// Don't log on cache miss
-			return []byte{}, time.Time{}, ErrNotCached
+			return []byte{}, time.Time{}, cache.ErrNotCached
 		} else {
 			c.logger.Log("err", errors.Wrap(err, "Fetching tag from memcache"))
 			return []byte{}, time.Time{}, err
@@ -133,7 +101,7 @@ func (c *memcacheClient) GetKey(k Keyer) ([]byte, time.Time, error) {
 	return cacheItem.Value[4:], time.Unix(int64(expiry), 0), nil
 }
 
-func (c *memcacheClient) SetKey(k Keyer, v []byte) error {
+func (c *memcacheClient) SetKey(k cache.Keyer, v []byte) error {
 	expiry := time.Now().Add(expiry).Unix()
 	expiryBytes := make([]byte, 4, 4)
 	binary.BigEndian.PutUint32(expiryBytes, uint32(expiry))
@@ -186,57 +154,4 @@ func (c *memcacheClient) updateMemcacheServers() error {
 	// guarantee best possible match between nodes.
 	sort.Strings(servers)
 	return c.serverList.SetServers(servers...)
-}
-
-// An interface to provide the key under which to store the data
-// Use the full path to image for the memcache key because there
-// might be duplicates from other registries
-type Keyer interface {
-	Key() string
-}
-
-type manifestKey struct {
-	fullRepositoryPath, reference string
-}
-
-func NewManifestKey(image image.CanonicalRef) Keyer {
-	return &manifestKey{image.CanonicalName().String(), image.Tag}
-}
-
-func (k *manifestKey) Key() string {
-	return strings.Join([]string{
-		"registryhistoryv3", // Just to version in case we need to change format later.
-		k.fullRepositoryPath,
-		k.reference,
-	}, "|")
-}
-
-type tagKey struct {
-	fullRepositoryPath string
-}
-
-func NewTagKey(id image.CanonicalName) Keyer {
-	return &tagKey{id.String()}
-}
-
-func (k *tagKey) Key() string {
-	return strings.Join([]string{
-		"registrytagsv3", // Just to version in case we need to change format later.
-		k.fullRepositoryPath,
-	}, "|")
-}
-
-type repoKey struct {
-	fullRepositoryPath string
-}
-
-func NewRepositoryKey(repo image.CanonicalName) Keyer {
-	return &repoKey{repo.String()}
-}
-
-func (k *repoKey) Key() string {
-	return strings.Join([]string{
-		"registryrepov3",
-		k.fullRepositoryPath,
-	}, "|")
 }
