+++ registry/cache/memcached/memcached_test.go
@@ -1,6 +1,5 @@
 // +build integration
-
-package cache
+package memcached
 
 import (
 	"flag"
