+++ registry/doc.go
@@ -1,0 +1,5 @@
+/*
+This package has types for dealing with image registries (e.g.,
+quay.io, DockerHub, Google Container Registry, ..).
+*/
+package registry
